requirements.txt
configobj==4.7.2
jabberbot==0.15
xmpppy==0.5.0rc1